ao_django/requirements.txt
channels
channels_redis
django-crispy-forms
django-invitations
social-auth-app-django
djangorestframework
django-tables2
django-datatable-view
tablib
bokeh